Fixed OBB render broken by Angelica
