+++ src/main/java/su/sergiusonesimus/metaworlds/zmixin/mixins/minecraft/client/renderer/MixinRenderGlobalVanilla.java
@@ -30,6 +30,7 @@ import net.minecraft.entity.EntityLiving;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.init.Blocks;
 import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.AxisAlignedBB;
 import net.minecraft.util.IIcon;
 import net.minecraft.util.MathHelper;
 import net.minecraft.util.Vec3;
@@ -43,13 +44,16 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.At.Shift;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.ModifyArgs;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
+import org.spongepowered.asm.mixin.injection.invoke.arg.Args;
 
 import com.llamalad7.mixinextras.sugar.Local;
 
 import su.sergiusonesimus.metaworlds.api.SubWorld;
 import su.sergiusonesimus.metaworlds.client.multiplayer.SubWorldClient;
+import su.sergiusonesimus.metaworlds.util.OrientedBB;
 import su.sergiusonesimus.metaworlds.zmixin.interfaces.minecraft.client.renderer.IMixinRenderGlobalVanilla;
 import su.sergiusonesimus.metaworlds.zmixin.interfaces.minecraft.client.renderer.IMixinRenderList;
 import su.sergiusonesimus.metaworlds.zmixin.interfaces.minecraft.util.IMixinAxisAlignedBB;
@@ -1189,6 +1193,59 @@ public class MixinRenderGlobalVanilla implements IMixinRenderGlobalVanilla {
         }
 
         ++this.frustumCheckOffset;
+    }
+
+    // drawOutlinedBoundingBox
+
+    private static AxisAlignedBB storedAABB;
+    private static OrientedBB storedOBB;
+
+    @Inject(method = "drawOutlinedBoundingBox", at = @At(value = "HEAD"))
+    private static void storeVariables(AxisAlignedBB aabb, int color, CallbackInfo ci) {
+        storedAABB = aabb;
+        storedOBB = ((IMixinAxisAlignedBB) aabb).getOrientedBB();
+    }
+
+    @ModifyArgs(
+        method = "drawOutlinedBoundingBox",
+        at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/Tessellator;addVertex(DDD)V"))
+    private static void modifyAddVertex(Args args) {
+        double x = args.get(0);
+        double y = args.get(1);
+        double z = args.get(2);
+        int vertexIndex;
+        if (y == storedAABB.minY) {
+            if (z == storedAABB.minZ) {
+                if (x == storedAABB.minX) {
+                    vertexIndex = 0;
+                } else {
+                    vertexIndex = 1;
+                }
+            } else {
+                if (x == storedAABB.minX) {
+                    vertexIndex = 2;
+                } else {
+                    vertexIndex = 3;
+                }
+            }
+        } else {
+            if (z == storedAABB.minZ) {
+                if (x == storedAABB.minX) {
+                    vertexIndex = 4;
+                } else {
+                    vertexIndex = 5;
+                }
+            } else {
+                if (x == storedAABB.minX) {
+                    vertexIndex = 6;
+                } else {
+                    vertexIndex = 7;
+                }
+            }
+        }
+        args.set(0, storedOBB.getX(vertexIndex));
+        args.set(1, storedOBB.getY(vertexIndex));
+        args.set(2, storedOBB.getZ(vertexIndex));
     }
 
 }
