+++ src/main/resources/mcmod.info
@@ -20,6 +20,7 @@
         	"before:McMultipart",
         	"before:HardcoreEnderExpansion",
         	"before:littletiles@[1.6.18,)",
-        	"before:backhand" ]
+        	"before:backhand",
+        	"before:angelica@[2.1.4,)" ]
     }
 ]
